Correction pour 42FileChecker

+++ Libft/ft_strcpy.c
@@ -6,7 +6,7 @@
 /*   By: wgulista <wgulista@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/26 18:26:34 by wgulista          #+#    #+#             */
-/*   Updated: 2015/11/26 18:28:23 by wgulista         ###   ########.fr       */
+/*   Updated: 2016/01/01 16:45:45 by wgulista         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strcpy(char *dest, const char *src)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (src[i] != '\0')
